requirements.txt
ImageHash==4.3.1
numpy==2.0.0
opencv-python-headless==4.10.0.84
pillow==10.4.0
PyWavelets==1.6.0
scipy==1.14.0
